requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
anthropic>=0.30.0
mcp>=1.0.0
python-dotenv>=1.0.0
